this is my menu page
I wanna do cube photo and descriptions on the side. I want to have a button that allows me to add or subtract the amount and one that allows me to move the amount of items into the pay area. I also want to add a pay button on everything.